% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDoubleAssignment.R
\name{removeDoubleAssignment}
\alias{removeDoubleAssignment}
\title{remove any users with double assignment events from experiment data}
\usage{
removeDoubleAssignment(raw_data, id_variable, sample_variable, control_value)
}
\description{
remove any users with double assignment events from experiment data
}
\examples{
removeDoubleAssignment(raw_test_data,user_id,sample,'is control')
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
