% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseBayesPropDaily.R
\name{summariseBayesPropDaily}
\alias{summariseBayesPropDaily}
\title{Takes raw data a constructs a daily time series from the date}
\usage{
summariseBayesPropDaily(daily_data, sig_level)
}
\description{
Takes raw data a constructs a daily time series from the date
}
\examples{
summariseBayesPropDaily()
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
