% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformData.R
\name{transformData}
\alias{transformData}
\title{Takes raw experiment data and transforms it into a format that can easily pass through all of the other functions downstream}
\usage{
transformData(
  raw_data,
  unique_id,
  numerator,
  denominator,
  sample_column,
  control_value,
  date_column
)
}
\description{
Takes raw experiment data and transforms it into a format that can easily pass through all of the other functions downstream
}
\examples{
transformData(test_data,user_id,value2,value,sample,'is control',date)
}
\keyword{ab}
\keyword{metric}
\keyword{test,}
\keyword{transform,}
