% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeBayesianPropDaily.R
\name{visualizeBayesianPropDaily}
\alias{visualizeBayesianPropDaily}
\title{This takes results from summariseMetricDaily and generates time-series view with cumulative daily totals and credible intervals.}
\usage{
visualizeBayesianPropDaily(daily_bayes)
}
\description{
This takes results from summariseMetricDaily and generates time-series view with cumulative daily totals and credible intervals.
}
\examples{
visualizeBayesianPropDaily(bayes_daily)
}
\keyword{ab}
\keyword{summarise}
\keyword{test,}
